Begin implementation of ANY card game.
The current implementation defines a Gambler class (player's and bankers
are both gamblers by definition). This Gambler class has a hand of
Cards, an integer representing the line of the save file on which player
data would be saved, and a takeCards method.

TODO: FIXME: the takeCards method is currently incorrect according to
the rules of Punto Banco. There are cases where less than two cards are
to be taken by the banker, and possibly the player. The method must have
a parameter defining the number of cards to draw from the CardDeck.