Override the Gambler.equals(Object o) method
This finally resolves the issue of not being able to use indexOf to find a
Gambler within the active, in-memory database (an array list of Gamblers) to
retrieve data from disk, or the modified in-memory database before it is dumped
back to disk for the next instance of the application.

Thanks to Taylor (CAMRU executive) and an online resource for instructing on how
to properly override the equals method.